relocate constants.py to the place a competent dev might have originally put it
