+++ src/pds_doi_service/core/util/config_parser.py
@@ -20,7 +20,7 @@ from os.path import abspath
 from os.path import dirname
 from os.path import join
 
-from constants import PROJECT_ROOT_DIR
+from pds_doi_service.constants import PROJECT_ROOT_DIR
 from pds_doi_service.core.util.logging import get_logger
 from pkg_resources import resource_filename
